Split ENSToken handlers into seperate file
