+++ subgraphs/ens-governance/src/ens-governor.ts
@@ -208,122 +208,3 @@ export function handleVoteCast(event: VoteCast): void {
   voter.numberVotes = voter.numberVotes + 1;
   voter.save();
 }
-
-// DelegateChanged(indexed address,indexed address,indexed address)
-export function handleDelegateChanged(event: DelegateChanged): void {
-  let tokenHolder = getTokenHolder(event.params.delegator.toHexString());
-  let previousDelegate = getDelegate(event.params.fromDelegate.toHexString());
-  let newDelegate = getDelegate(event.params.toDelegate.toHexString());
-
-  tokenHolder.delegate = newDelegate.id;
-  tokenHolder.save();
-
-  previousDelegate.tokenHoldersRepresentedAmount =
-    previousDelegate.tokenHoldersRepresentedAmount - 1;
-  previousDelegate.save();
-
-  newDelegate.tokenHoldersRepresentedAmount =
-    newDelegate.tokenHoldersRepresentedAmount + 1;
-  newDelegate.save();
-}
-
-// DelegateVotesChanged(indexed address,uint256,uint256)
-// Called in succession to the above DelegateChanged event
-export function handleDelegateVotesChanged(event: DelegateVotesChanged): void {
-  const delegateAddress = event.params.delegate;
-  const previousBalance = event.params.previousBalance;
-  const newBalance = event.params.newBalance;
-
-  let votesDifference = newBalance.minus(previousBalance);
-
-  let delegate = new Delegate(delegateAddress.toHexString());
-  delegate.delegatedVotesRaw = newBalance;
-  delegate.delegatedVotes = toDecimal(newBalance);
-  delegate.save();
-
-  // Update governance delegate count
-  let governance = getGovernance();
-  if (previousBalance == BIGINT_ZERO && newBalance > BIGINT_ZERO) {
-    governance.currentDelegates = governance.currentDelegates.plus(BIGINT_ONE);
-  }
-  if (newBalance == BIGINT_ZERO) {
-    governance.currentDelegates = governance.currentDelegates.minus(BIGINT_ONE);
-  }
-  governance.delegatedVotesRaw = governance.delegatedVotesRaw.plus(
-    votesDifference
-  );
-  governance.delegatedVotes = toDecimal(governance.delegatedVotesRaw);
-  governance.save();
-}
-
-// Transfer(indexed address,indexed address,uint256)
-export function handleTransfer(event: Transfer): void {
-  const from = event.params.from;
-  const to = event.params.to;
-  const value = event.params.value;
-
-  let fromHolder = getTokenHolder(from.toHexString());
-  let toHolder = getTokenHolder(to.toHexString());
-  let governance = getGovernance();
-
-  // Deduct from from holder balance + decrement gov token holders
-  // if holder now owns 0 or increment gov token holders if new holder
-  if (from.toHexString() != ZERO_ADDRESS) {
-    let fromHolderPreviousBalance = fromHolder.tokenBalanceRaw;
-    fromHolder.tokenBalanceRaw = fromHolder.tokenBalanceRaw.minus(value);
-    fromHolder.tokenBalance = toDecimal(fromHolder.tokenBalanceRaw);
-
-    if (fromHolder.tokenBalanceRaw < BIGINT_ZERO) {
-      log.error("Negative balance on holder {} with balance {}", [
-        fromHolder.id,
-        fromHolder.tokenBalanceRaw.toString(),
-      ]);
-    }
-
-    if (
-      fromHolder.tokenBalanceRaw == BIGINT_ZERO &&
-      fromHolderPreviousBalance > BIGINT_ZERO
-    ) {
-      governance.currentTokenHolders = governance.currentTokenHolders.minus(
-        BIGINT_ONE
-      );
-      governance.save();
-    } else if (
-      fromHolder.tokenBalanceRaw > BIGINT_ZERO &&
-      fromHolderPreviousBalance == BIGINT_ZERO
-    ) {
-      governance.currentTokenHolders = governance.currentTokenHolders.plus(
-        BIGINT_ONE
-      );
-      governance.save();
-    }
-
-    fromHolder.save();
-  }
-
-  // Increment to holder balance and total tokens ever held
-  let toHolderPreviousBalance = toHolder.tokenBalanceRaw;
-  toHolder.tokenBalanceRaw = toHolder.tokenBalanceRaw.plus(value);
-  toHolder.tokenBalance = toDecimal(toHolder.tokenBalanceRaw);
-  toHolder.totalTokensHeldRaw = toHolder.totalTokensHeldRaw.plus(value);
-  toHolder.totalTokensHeld = toDecimal(toHolder.totalTokensHeldRaw);
-
-  if (
-    toHolder.tokenBalanceRaw == BIGINT_ZERO &&
-    toHolderPreviousBalance > BIGINT_ZERO
-  ) {
-    governance.currentTokenHolders = governance.currentTokenHolders.minus(
-      BIGINT_ONE
-    );
-    governance.save();
-  } else if (
-    toHolder.tokenBalanceRaw > BIGINT_ZERO &&
-    toHolderPreviousBalance == BIGINT_ZERO
-  ) {
-    governance.currentTokenHolders = governance.currentTokenHolders.plus(
-      BIGINT_ONE
-    );
-    governance.save();
-  }
-  toHolder.save();
-}
